added hdfs tests
